Разделил проект на модули
