+++ js/util.js
@@ -1,0 +1,27 @@
+//Вспомогательные функции
+
+const getRandomInteger = (a, b) => {
+  const lower = Math.ceil(Math.min(a, b));
+  const upper = Math.floor(Math.max(a, b));
+  const result = Math.random() * (upper - lower + 1) + lower;
+  return Math.floor(result);
+};
+
+const generateRandomIndex = (a, b) => {
+  const indexNumbers = [];
+  return () => {
+    let currentIndex = getRandomInteger(a, b);
+    if (indexNumbers.length === Math.floor(Math.max(a, b))) {
+      return '';
+    }
+    while (indexNumbers.includes(currentIndex)) {
+      currentIndex = getRandomInteger(a, b);
+    }
+    indexNumbers.push(currentIndex);
+    return currentIndex;
+  };
+};
+
+const getElementOfArray = (arr) => arr[getRandomInteger(0, arr.length - 1)];
+
+export { getElementOfArray, generateRandomIndex, getRandomInteger };
